wip: improve emoji react
